Split DataServlet into separate Servlet Actions
