+++ portfolio/src/main/webapp/index.html
@@ -33,7 +33,7 @@
           <!-- Comment Container -->
           <div id="comments_container">
             <div id="comment_section_wrapper">
-              <form action="/comments" method="POST">
+              <form action="/new-comment" method="POST">
                 <h2 class="page_headers">Leave a comment!</h2>
                 <textarea 
                   type="text" 
@@ -54,6 +54,9 @@
             </div>
             <br/><br/>
             <div id="comment-display-container"></div>
+            <form action="/new-comment" method="POST">
+              <input type="submit" alt="Click to submit your comment" id="comment_submit"/>
+            </form>
           </div>
         </div>
         <div id="content">
